requirements.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
earthengine-api==0.1.240
Fiona==1.8.17
future==0.18.2
gast==0.3.3
GeoAlchemy2==0.8.4
geopandas==0.8.1
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.1
h5py==2.10.0
httplib2==0.18.1
httplib2shim==0.0.3
idna==2.10
importlib-metadata==2.0.0
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.2
matplotlib==3.3.2
munch==2.5.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.3
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pvlib==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygeos==0.8
pyparsing==2.4.7
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.3.20
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.1
